+++ src/components/speakers/speaker-card.tsx
