+++ src/components/speakers/speaker.interface.ts
