+++ src/components/sponsors/partner-card.tsx
