+++ src/components/sponsors/sponsor-card.tsx
